fix: rename delete handlers file
