+++ frontend/src/app-components/dialogs/utils/deleteHandlers.ts
