+++ frontend/src/components/content-types/index.tsx
@@ -12,7 +12,7 @@ import { Button, Grid, Paper } from "@mui/material";
 import { useRouter } from "next/router";
 
 import { DeleteDialog } from "@/app-components/dialogs";
-import { deleteCallbackHandler } from "@/app-components/dialogs/utils/deleteHandles";
+import { deleteCallbackHandler } from "@/app-components/dialogs/utils/deleteHandlers";
 import { FilterTextfield } from "@/app-components/inputs/FilterTextfield";
 import {
   ActionColumnLabel,
